app/core/config.py
import os

from dotenv import load_dotenv
from pydantic.v1 import BaseSettings

load_dotenv()

class Settings(BaseSettings):
    DATABASE_URL: str = os.getenv("DATABASE_URL")
    secret_key: str = os.getenv("SECRET_KEY")
    algorithm: str = os.getenv("ALGORITHM")
    access_token_expire_minutes: int = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES"))
    login_max_attempts: int = int(os.getenv("MAX_ATTEMPTS"))
    rate_block_time: int = int(os.getenv("BLOCK_TIME"))
settings = Settings()
